ParamedicApp: app creation methods extracted from paramedic.js
